Moves apps separate packages
